requirements.txt
gradio-rangeslider==0.0.8
gradio==5.44.1
librosa==0.11.0
numpy==2.2.1
onnx==1.20.1
onnxruntime-gpu==1.24.3
opencv-python==4.13.0.92
tqdm==4.67.3
scipy==1.17.1
